catalystic-backend/requirements.txt
requests
python-dotenv
supabase 